src/requirements.txt
exceptiongroup==1.2.2
iniconfig==2.0.0
lxml==5.3.0
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
pytest==8.3.3
python-docx==1.1.2
python-dotenv==1.0.1
tomli==2.0.2
typing_extensions==4.12.2
